requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
pdfplumber>=0.10.2
pymupdf>=1.23.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
chromadb>=0.4.18
langchain-openai>=0.0.2
langchain-community>=0.0.13
python-dotenv>=1.0.0
requests>=2.31.0
flask>=2.3.0
werkzeug>=2.3.0
